colorspace: remove DCDM chromatic adaptation
The colors are encoded in XYZ with EE white point, but this is only the
encoding part. Displayed white point is at code poit X'=3794 Y'=3960
Z'=3890. Reference projector is expected to produce DCI white for those
values.

Although SMPTE standard does not enforce specific white point. For
example ACES uses D60 instead. But the workflow for reference projector
to be DCI compliant and it is on encoders to ensure the content looks
good.

So this is what we do. Decode DCDM X'Y'Z into DCI-P3 RGB.