meson: add glslang lib for 15.0.0 linking
Glslang 15.0.0 moved some code around, add also linking to glslang,
while this is not needed for older versions, it will still work.

Unfortunately CMake config embedded in distributions is not usable
without installing additional package like glslang-tools, because at
least Ubuntu splits it. On Arch it would work, but generally the CMake
config require cmake binary to work also, so let's keep it as-is for
now.