shaders: add #line before generated function starts
This fixes errors with coverage intrumentation enabled:

error: function starts on a higher line number than it ends
[-Werror=coverage-invalid-line-number]

It's still not perfect, as we can have colisions of the line numbers
between call site and function site. Probably would be better to
generate functions in separate file and #include it. Then the line
numbering will refer to different file name.

But for now this fixes the build with GCC 13.