renderer: fixup component_mapping during plane merge
It may happen that target format during merge has different sampling
order than the source format, for example rg8 merged with r8 into bgrx8.
In this case merge shader will implicitly reorder components and since
component_mapping is not connected to sample_order, it has to be updated
to match the new reality. Note that we don't apply component_mapping
during merge, it will be applied at the end of read_image pass.

Fixes: https://github.com/mpv-player/mpv/issues/16208