vulkan/context: remove KHR/EXT suffixes when promoting extention
Aliases (function names) are provided by corresponding entity which
introduced it. We were promoting extension to core, basically skipping
loading given extension, but still trying to load alias from that
extension.

We can read in documentation that:
A function pointer for a given alias can only be retrieved if the
extension or version that introduced that alias is supported and
enabled, irrespective of whether any other alias is available.

According to vkGetDeviceProcAddr behavior, loading functions from
disable extensions returns NULL. Unfortunately that's not the case in
fully CTS compliant Vulkan 1.4 implementations, so work around this
issue, by just not trying to get pointer to those functions. We already
know for sure that they shouldn't be needed and even available.