shaders/colorspace: don't apply alpha mode during color encode/decode
This is the wrong place to apply the alpha mode, because it leads to double
applications when the renderer is already changing the alpha mode. It's better
to let the renderer handle this logic, since it's the only place that actually
keeps track of the current state of the image.

pl_color_repr_encode/decode merely _assumed_ that the image was straight alpha
going in.