colorspace: fix math in pl_primaries_compatible()
This function is supposed to work by checking if each primary in
colorspace a is closer to the corresponding primary in colorspace b than
to that colorspace's other primaries. However, that's not what it
actually does for the blue and green primaries: instead of computing
"GR", the distance from a's green to b's red, it reuses "RG", the
distance from a's red to b's green, which is not the same. The same
issue applies to "BR" and "BG", which should both be computed but
aren't.

See this commit's Merge Request[1] for diagrams illustrating the issue.

[1] https://code.videolan.org/videolan/libplacebo/-/merge_requests/702