vulkan/formats: add support for VK_KHR_format_feature_flags_2
Because we can't easily tell whether or not the extension is loaded, just guard
it by vulkan version 1.3 instead. It shouldn't matter, as the extension on
its own doesn't do anything, but we still load it for downstream dependants.

This required special handling for the DRM properties, because they magically
upgrade themselves to the new extension iff it is available. As an aside, we
are no longer at risk of truncating the modifier list.