vulkan/gpu: add VK_EXT_host_image_copy support
This is substantially more performant than importing a pointer as a
host-mapped buffer and then issuing an asynchronous upload on the GPU,
although it comes at the cost of being synchronous and hence not leveraging
the async DMA engine.

I found it to be significantly slower for downloads as opposed to uploads,
though, even when only enabling them opportunistically on otherwise idle
textures. So stick with the upload-only path for now.