shaders/colorspace: re-add correct alpha normalization
This was broken in the past, and also not handled in a very elegant way by
the renderer. During the process of fixing the renderer's alpha handling,
this logic was removed in 0067206ff without a replacement. However, that
is also not the correct fix, as the documentation clearly states that the
alpha mode is applied by these functions, and other places in the renderer
depended on this behavior.

Instead, the correct fix to these functions is to re-add the alpha mode
application, but at the correct place in the pipeline, which is on the
RGBA *ingest*, rather than the encoded color. (Basically the opposite place
of what the previous logic was doing)

Fixes: 0067206ff880341fafa621a41d99d63fd81e3e8e
Fixes: https://code.videolan.org/videolan/libplacebo/-/issues/336
Fixes: https://github.com/haasn/libplacebo/issues/332
Fixes: https://github.com/mpv-player/mpv/issues/16396
Closes: https://code.videolan.org/videolan/libplacebo/-/merge_requests/709