vulkan/swapchain: prefer VK_COLOR_SPACE_PASS_THROUGH_EXT
If neither the primaries nor the transfer function match any of the
color spaces, before to this change, the first one on the list with the
best score would be selected.

Now a small score increase is given to the pass through color space,
which is a better choice, especially when setting metadata.

This fixes a case where, for a BT.2020/PQ request, the sRGB color space was
selected instead of the pass-through one. This happened because both had the
same score and sRGB appeared earlier in the array.