vulkan/gpu_tex: fix validation error on pure layout transition
When transitioning an image layout (e.g. with pl_vulkan_hold_ex), we emit
a pipeline barriar with the destination stage set to NONE, as recommended
by the Vulkan specification.

However, when the source stage is also set to NONE, this introduces a race
condition, at least according to the validation layers. Adding at least some
access flag in this case seems to solve this issue.