vulkan/swapchain: guard against out-of-order acquisition
Sometimes, the swapchain likes to acquire images out of order, without
ever giving us back the earliest one. This can lead to a situation
where we re-use a semaphore that was last used to present a past image,
for which we have no proof of completion because the swapchain never
let us re-acquire it.

The simplest solution is to decouple the choice of semaphore used for
acquisition and presentation. For the presentation side, this guarantees that
the semaphore is not currently in use, otherwise the swapchain would not
have given us back that image.

For the acquisition side, by just allocating one extra semaphore, we should
be able to guarantee that the least recently used semaphore is always not
in use, even if every single frame is currently queued. I'm not 100%
confident this is actually a sufficient condition, but testing in practice
I was at least not able to trigger any issues anymore.