utils/libav_internal: map 10X6 / 12X4 formats to 16 bit equivalents
These formats are often less supported. If we create the image with the
mutable format bit, we might as well reinterpret them to a more supported
format.

I would love to do this internally inside pl_vulkan_wrap(), but we don't have
access to the corresponding pl_bit_encoding struct inside there, so we're
forced to handle it in generic code.

Fixes issues with rendering to p010le textures on nvidia.