renderer: add support for clearing with blurred copy of image
Uses a dual kawase blur filter for a great mix of economy and quality.
In theory, we could mask out the portions of the blur pass that we don't
need in the final result, but this is quite nontrivial to compute. For now,
we just avoid re-drawing that during the final bicubic border blend pass.

This is not quite an ideal Gaussian filter, so the transitions between
different iteration counts can be kind of rough, but the overall effect still
works well enough.