Revert "vulkan/swapchain: check if host has color management actually enabled"
This reverts commit 51c7d2c1942d988be6f155dc0b57770550c8ff48.

This is completely dead code because colorspaces added by
VK_EXT_swapchain_colorspace will never be advertised to us if the host
doesn't advertise support for the necessary primaries/transfer
functions.

If the host is claiming support for things it doesn't indeed support,
then it's a bug there and shouldn't be worked around in libplacebo.

Moreover, it seems this heuristic for checking for color management
support completely fails on Nvidia* at least, so let's not have this in
libplacebo.

* https://projects.blender.org/blender/blender/pulls/133159#issuecomment-1399531