vulkan/command: tie semaphores to queues, not commands
There is no reason for us to have semaphores on a so granular basis - we
only ever need to keep track of commands executed on a given queue, which
will always be executed in submission order already.

Unfortunately, this would break if we end up trying to record two commands
on the same queue at the same time, since the semaphore value would end
up being inconsistent with the submission order. Fortunately, this is never
a case we can currently hit, because all recorded commands are already
locked by the pl_gpu's internal state.

Still, add a safety assertion just to be sure.