include/utils/libav: add sanity check to pl_map_avframe_drm()
When called on unsupported formats, `out->num_planes` may be zero. This is
normally caught downstream by the format mapping check, but that happens
*after* we attempt accessing the texture inside the same function.

Just add an early sanity check to be on the safe side. Even if we end up
adding a second safety check to the usage site, it's better to have this
check here in any case, to make the function more self-contained.

Fixes: https://github.com/haasn/libplacebo/issues/327